filter_kernels: Separate src and dst radius and remove inv_scale.
When downsampling, it is necessary to widen the radius of the
filter to include more samples.  However, the shape and scale of the
underlying windowed filter function should not change when calculating
the weight coefficients for samples.

The existing code uses a single radius value that corresponds to
the source sample radius and tries to correct this value to the
destination (windowed filter) radius/coordinate system as needed.

This approach is tricky to reason about and has a few spots where
the proper radius value is not used.

This commit instead removes the correction factor and just directly
calculates a source (sample) radius and a destination
(window filter function) radius.  This makes it much easier to
determine whether the proper value is used at any given location
in the code.